requirements.txt
streamlit
easyocr
pytesseract
paddleocr
transformers
torch
pillow
google-generativeai
opencv-python
numpy
pandas
